Feature: define a custom user-class
Identify this iPXE ROM as "BlueBanquise". It allows to discriminate this
ROM from other iPXE ROM that may not comply with BlueBanquise's PXE
chain. Comply with RFC 3004 (https://tools.ietf.org/html/rfc3004).

Note: RFC 3004 defines the DHCP user class as a set of length-value
tuples, but iPXE treats it as a string. You can choose to manually
construct a value which conforms to RFC 3004 using the set command. This
is a long-standing bug both in iPXE and in the reference DHCP server
implementation (ISC dhcpd). There is a substantial amount of
documentation which suggests checking option 77 for the value 'iPXE' to
identify iPXE clients, and using 'set user-class <text>' to specify a
custom user class. To avoid breaking existing setups, iPXE will continue
to provide option 77 as a plain string for DHCPv4.
(Source: http://ipxe.org/cfg/user-class)